requirements.txt
aiohttp
pyrogram
tgcrypto
asyncpg